requirements.txt
# Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Computer Vision & ML
opencv-python==4.8.1.78
mediapipe==0.10.14
numpy>=1.26.0
pillow==10.1.0
scikit-learn==1.3.2
joblib==1.3.2

# Deep Learning (for temporal gesture detection)
torch==2.9.1
torchvision==0.24.1

# AI Ops & MLflow
mlflow==2.22.4
prometheus-client==0.19.0

# Data handling
pydantic==2.5.0
pydantic-settings==2.1.0

# Utilities
python-dotenv==1.0.0
requests==2.32.5
